+++ handlers/routes.go
@@ -112,7 +112,6 @@ func Router() *mux.Router {
 	api.Handle("/api/services/{id}/failure_data", cached("30s", "application/json", apiServiceFailureDataHandler)).Methods("GET")
 	api.Handle("/api/services/{id}/ping_data", cached("30s", "application/json", apiServicePingDataHandler)).Methods("GET")
 	api.Handle("/api/services/{id}/uptime_data", http.HandlerFunc(apiServiceTimeDataHandler)).Methods("GET")
-	//api.Handle("/api/services/{id}/heatmap", cached("30s", "application/json", apiServiceHeatmapHandler)).Methods("GET")
 
 	// API INCIDENTS Routes
 	api.Handle("/api/services/{id}/incidents", http.HandlerFunc(apiServiceIncidentsHandler)).Methods("GET")
@@ -162,8 +161,8 @@ func Router() *mux.Router {
 	r.Handle("/health", http.HandlerFunc(healthCheckHandler))
 	r.Handle("/oauth/{provider}", http.HandlerFunc(oauthHandler))
 	r.Handle("/.well-known/", http.StripPrefix("/.well-known/", http.FileServer(http.Dir(dir+"/.well-known"))))
-
 	r.NotFoundHandler = http.HandlerFunc(error404Handler)
+	r.Handle("/", http.HandlerFunc(indexHandler))
 	return r
 }