+++ types/configs/methods.go
@@ -9,17 +9,3 @@ func (d *DbConfig) Save(directory string) error {
 	}
 	return nil
 }
-
-// defaultPort accepts a database type and returns its default port
-func defaultPort(db string) int {
-	switch db {
-	case "mysql":
-		return 3306
-	case "postgres":
-		return 5432
-	case "mssql":
-		return 1433
-	default:
-		return 0
-	}
-}