+++ utils/utils_test.go
@@ -189,3 +189,18 @@ func TestHttpRequest(t *testing.T) {
 	assert.Equal(t, []byte("OK"), body)
 	assert.Equal(t, resp.StatusCode, 200)
 }
+
+func TestConfigLoad(t *testing.T) {
+	os.Setenv("DB_CONN", "sqlite")
+	InitCLI()
+	setDefaults()
+
+	s := Params.GetString
+	b := Params.GetBool
+
+	assert.Equal(t, "sqlite", s("DB_CONN"))
+	assert.Equal(t, Directory, s("STATPING_DIR"))
+	assert.True(t, b("SAMPLE_DATA"))
+	assert.False(t, b("DISABLE_LOGS"))
+	assert.False(t, b("ALLOW_REPORTS"))
+}