+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fixed postgres timestamp grouping
 - Added postman (newman) API testing
 - Added Viper and Cobra config/env parsing package
+- Added more golang tests
 
 # 0.90.27 (04-15-2020)
 - Fixed postgres database table creation process